+++ connector/src/test/java/it/eng/connector/integration/negotiation/ContractNegotiationRequestedIntegrationTest.java
@@ -7,42 +7,84 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.util.Arrays;
+import java.util.Collections;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.web.servlet.ResultActions;
 
+import it.eng.catalog.model.Catalog;
+import it.eng.catalog.model.Dataset;
+import it.eng.catalog.repository.CatalogRepository;
+import it.eng.catalog.repository.DatasetRepository;
+import it.eng.catalog.util.CatalogMockObjectUtil;
 import it.eng.connector.integration.BaseIntegrationTest;
 import it.eng.connector.util.TestUtil;
+import it.eng.negotiation.model.Action;
 import it.eng.negotiation.model.ContractNegotiation;
 import it.eng.negotiation.model.ContractNegotiationErrorMessage;
 import it.eng.negotiation.model.ContractNegotiationState;
 import it.eng.negotiation.model.ContractRequestMessage;
 import it.eng.negotiation.model.NegotiationMockObjectUtil;
 import it.eng.negotiation.model.Offer;
+import it.eng.negotiation.model.Permission;
 import it.eng.negotiation.serializer.NegotiationSerializer;
 
 public class ContractNegotiationRequestedIntegrationTest extends BaseIntegrationTest {
 // -> REQUESTED
 //	@PostMapping(path = "/request")
 	
+	@Autowired
+	private CatalogRepository catalogRepository;
+	@Autowired
+	private DatasetRepository datasetRepository;
+	
+	private Catalog catalog;
+	private Dataset dataset;
+	
+	@BeforeEach
+	public void populateCatalog() {
+		dataset = Dataset.Builder.newInstance()
+				.hasPolicy(Collections.singleton(CatalogMockObjectUtil.OFFER))
+				.build();
+		catalog = Catalog.Builder.newInstance()
+				.dataset(Collections.singleton(dataset))
+				.build();
+		
+		datasetRepository.save(dataset);
+		catalogRepository.save(catalog);
+	}
+	
+	@AfterEach
+	public void cleanup() {
+		datasetRepository.deleteAll();
+		catalogRepository.deleteAll();
+	}
+	
     @Test
     @WithUserDetails(TestUtil.CONNECTOR_USER)
     public void createNegotiation_success() throws Exception {
-    	
-    	//needs to match offer in initial.data
-    	Offer offer = Offer.Builder.newInstance()
-    			.id(offerID)
-    			.target(NegotiationMockObjectUtil.TARGET)
+    	//needs to match offer in catalog
+    	Permission permsission = Permission.Builder.newInstance()
+    			.action(Action.USE)
+    			.constraint(Arrays.asList(NegotiationMockObjectUtil.CONSTRAINT_COUNT_5))
+    			.build();
+    	String datasetOfferId = dataset.getHasPolicy().stream().findFirst().get().getId();
+    	Offer offerRequest = Offer.Builder.newInstance()
+    			.id(datasetOfferId)
+    			.target(dataset.getId())
     			.assigner(NegotiationMockObjectUtil.ASSIGNER)
-    			.permission(Arrays.asList(NegotiationMockObjectUtil.PERMISSION_COUNT_5))
+    			.permission(Arrays.asList(permsission))
     			.build();
     	
     	ContractRequestMessage contractRequestMessage = ContractRequestMessage.Builder.newInstance()
     			.callbackAddress(NegotiationMockObjectUtil.CALLBACK_ADDRESS)
     			.consumerPid(NegotiationMockObjectUtil.CONSUMER_PID)
-    			.offer(offer)
+    			.offer(offerRequest)
     			.build();
     	
     	final ResultActions result =
@@ -58,7 +100,8 @@ public class ContractNegotiationRequestedIntegrationTest extends BaseIntegration
     	assertNotNull(contractNegotiationRequested);
     	assertEquals(ContractNegotiationState.REQUESTED, contractNegotiationRequested.getState());
     	
-    	offerCheck(getContractNegotiationOverAPI(contractNegotiationRequested.getConsumerPid(), contractNegotiationRequested.getProviderPid()));
+    	offerCheck(getContractNegotiationOverAPI(contractNegotiationRequested.getConsumerPid(), 
+    			contractNegotiationRequested.getProviderPid()), CatalogMockObjectUtil.OFFER.getId());
     }
     
     @Test
@@ -133,7 +176,7 @@ public class ContractNegotiationRequestedIntegrationTest extends BaseIntegration
     public void createNegotiation_invalid_offer_constraint() throws Exception {
     	// offer with dateTime constraint - will not match with one in initial_data
     	Offer offer = Offer.Builder.newInstance()
-    			.id(offerID)
+    			.id("TO BE CHANGED")
     			.target(NegotiationMockObjectUtil.TARGET)
     			.assigner(NegotiationMockObjectUtil.ASSIGNER)
     			.permission(Arrays.asList(NegotiationMockObjectUtil.PERMISSION))